[REFACTOR][PYTHON] Consolidate derived_object into tvm.ir.utils
derived_object was duplicated byte-for-byte across
python/tvm/runtime/support.py and
python/tvm/s_tir/meta_schedule/utils.py. The function is not a runtime
feature and is used outside meta_schedule (tvm.relax, tvm.tirx).

Move the single canonical definition into python/tvm/ir/utils.py.
tvm.ir loads before tvm.tirx and tvm.s_tir, so eager top-level imports
work from every consumer without load-order shenanigans.

Rewrite all 25 caller imports (runtime/support importers, meta_schedule
production sites, tests). Delete python/tvm/runtime/support.py and
remove the duplicate from python/tvm/s_tir/meta_schedule/utils.py.
