+++ python/tvm/ir/utils.py
@@ -14,8 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-"""Runtime support infra of TVM."""
+"""Utilities shared across TVM IR packages."""
 
 from typing import TypeVar
