+++ python/tvm/s_tir/meta_schedule/__init__.py
@@ -53,5 +53,4 @@ from .task_scheduler import TaskScheduler
 from .tir_integration import tune_tir
 from .tune import tune_tasks
 from .tune_context import TuneContext
-from .utils import derived_object
 from .post_optimization import post_opt
