+++ python/tvm/s_tir/meta_schedule/builder/local_builder.py
@@ -25,12 +25,13 @@ from typing import Optional, Union
 from tvm_ffi import register_global_func
 
 from tvm.ir import IRModule
+from tvm.ir.utils import derived_object
 from tvm.runtime import Module, Tensor, load_param_dict, save_param_dict
 from tvm.support.popen_pool import MapResult, PopenPoolExecutor, StatusKind
 from tvm.target import Target
 
 from ..logging import get_logger
-from ..utils import cpu_count, derived_object, get_global_func_with_default_on_worker
+from ..utils import cpu_count, get_global_func_with_default_on_worker
 from .builder import BuilderInput, BuilderResult, PyBuilder
 
 logger = get_logger(__name__)  # pylint: disable=invalid-name
