+++ python/tvm/s_tir/meta_schedule/cost_model/mlp_model.py
@@ -32,6 +32,7 @@ import numpy as np  # type: ignore
 import torch  # type: ignore
 
 import tvm
+from tvm.ir.utils import derived_object
 from tvm.support.tar import tar, untar
 
 from ....runtime import Tensor
@@ -43,7 +44,7 @@ from ..logging import get_logger
 from ..runner import RunnerResult
 from ..search_strategy import MeasureCandidate
 from ..tune_context import TuneContext
-from ..utils import derived_object, shash2hex
+from ..utils import shash2hex
 
 logger = get_logger("mlp_model")  # pylint: disable=invalid-name
