+++ python/tvm/s_tir/meta_schedule/runner/local_runner.py
@@ -22,12 +22,13 @@ from collections.abc import Callable
 from contextlib import contextmanager
 
 import tvm
+from tvm.ir.utils import derived_object
 from tvm.support.popen_pool import PopenPoolExecutor
 
 from ....runtime import Device, Module
 from ..logging import get_logger
 from ..profiler import Profiler
-from ..utils import derived_object, get_global_func_with_default_on_worker
+from ..utils import get_global_func_with_default_on_worker
 from .config import EvaluatorConfig
 from .runner import PyRunner, PyRunnerFuture, RunnerFuture, RunnerInput, RunnerResult
 from .utils import (
