+++ tests/python/s_tir/meta_schedule/test_meta_schedule_search_strategy.py
@@ -23,9 +23,9 @@ import pytest
 
 import tvm
 import tvm.testing
+from tvm.ir.utils import derived_object
 from tvm.s_tir import meta_schedule as ms
 from tvm.s_tir.meta_schedule.testing.dummy_object import DummyMutator
-from tvm.s_tir.meta_schedule.utils import derived_object
 from tvm.s_tir.schedule import Schedule, Trace
 from tvm.script import tirx as T
